Extension of Loading Animation!
The loading animation now starts from the old percentage value instead of
0 for smoother user experience.

Now it's even possible to provide different values for the desired start
and end position of the arc. E.g. an animation can be done from 37% to 86%

All that by taking into account the current selected interpolator! Because
an animation from 0% to 25% might last a different amount of time than an
animation from 75% to 100% although the values are both 25% appart.